requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
cssselect==1.1.0
Django==3.0.4
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
djangorestframework==3.11.0
dnspython==2.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
idna==2.9
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
mongoengine==0.23.1
newspaper3k==0.2.8
nltk==3.6.3
Pillow==8.3.2
pymongo==3.12.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2019.3
PyYAML==5.4.1
regex==2021.9.24
requests==2.23.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.0
sqlparse==0.3.1
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
urllib3==1.25.8
